+++ core/clients/db/filter/filter_pagination.go
@@ -19,13 +19,13 @@ func (f *Input) applyPagination(db *gorm.DB) *gorm.DB {
 	}
 
 	db = db.Scopes(
-		scope.Paginate(*f.PageNr, *f.NrOfItems, *f.maxNrOfItems),
+		scope.Paginate(int(*f.PageNr), int(*f.NrOfItems), int(*f.maxNrOfItems)),
 	)
 	return db
 }
 
 // SetNrOfItems -> this is the requested nr of items!
-func (f *Input) SetNrOfItems(nrOfItems int) *Input {
+func (f *Input) SetNrOfItems(nrOfItems int64) *Input {
 	f.NrOfItems = &nrOfItems
 	return f
 }
@@ -33,12 +33,12 @@ func (f *Input) SetNrOfItems(nrOfItems int) *Input {
 // SetMaxNrOfItems -> this is the max allowed value that can be set when requesting... or
 // the max value that will be taken in case of if NrOfItems is higher than this limit!
 // -1 can also be set! -> this means it's ALL or unlimited!
-func (f *Input) SetMaxNrOfItems(maxNrOfItems int) *Input {
+func (f *Input) SetMaxNrOfItems(maxNrOfItems int64) *Input {
 	f.maxNrOfItems = &maxNrOfItems
 	return f
 }
 
-func (f *Input) SetPageNr(pageNr int) *Input {
+func (f *Input) SetPageNr(pageNr int64) *Input {
 	f.PageNr = &pageNr
 	return f
 }
